Rework semantic analysis again
The old architecture although efficient is not really extendable -
adding new checks would require changes in a lot of existing
functions. Reworking it to make it easy to add new checks. Now each
checker or validator traverses the AST again, but this is an RPC so
there shouldn't be thousands of interfaces, right? Right?